api/requirements.txt
Flask
gunicorn
python-dotenv
markdown
requests
neo4j
spacy
google-cloud-aiplatform
google-generativeai
google-auth
google-api-core
flask-cors